refactor: Se reorganizaron los enums
- Se eliminó el archivo `enums.py` y se movieron las enumeraciones a `status_enum.py` y `role_moodle_enum.py`.
- Se actualizaron las rutas de importación en varios controladores y servicios.
