+++ app/modules/auth/routes.py
@@ -3,10 +3,10 @@ from fastapi import APIRouter, Depends, Query
 from app.core.database import get_db
 from app.modules.auth.controllers.get_user_info import GetUserInfoController
 from app.shared.dependecies.get_current_user import get_current_user
-from app.shared.enums.enums import AccountRoleEnum
 from app.shared.enums.institutes_enum import InstitutesEnum
+from app.shared.enums.role_enum import AccountRoleEnum
 
-from ...shared.enums.enums import AccountStatusEnum
+from ...shared.enums.status_enum import AccountStatusEnum
 from .controllers.change_status import ChangeStatusController
 from .controllers.create_account import CreateAccountController
 from .controllers.list_account_requests import ListAccountRequestsController
