+++ app/shared/enums/role_enum.py
@@ -1,12 +1,6 @@
 from enum import Enum
 
 
-class RoleEnum(str, Enum):
-    MANAGER = 1
-    COURSE_CREATOR = 2
-    EDITING_TEACHER = 3
-    TEACHER = 4
-    STUDENT = 5
-    GUEST = 6
-    USER = 7
-    FRONTPAGE = 8
+class AccountRoleEnum(str, Enum):
+    ALUMNO = "alumno"
+    DOCENTE = "docente"
