+++ app/shared/enums/status_enum.py
@@ -6,9 +6,3 @@ class AccountStatusEnum(Enum):
     APPROVED = "approved"
     REJECTED = "rejected"
     CREATED = "created"
-
-
-# Se agrego lo roles
-class AccountRoleEnum(str, Enum):
-    ALUMNO = "alumno"
-    DOCENTE = "docente"
